채팅 목록 클릭 시 색 변화 추가, 자동 스크롤 내리기, 엔터로 전송, Chat관련 키 데이터 타입 Long으로 변경

+++ src/main/java/com/example/panda/entity/ChatEntity.java
@@ -14,7 +14,7 @@ import java.time.LocalDateTime;
 public class ChatEntity {
     @Id // pk
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int chat_id;
+    private Long chat_id;
 
     @ManyToOne
     @JoinColumn(name="room_id")
